refact(Statistic): move balance components to separate folder
