+++ src/components/statistic/balance/page.tsx
@@ -3,35 +3,20 @@ import * as React from 'react';
 import { PageHeader } from 'react-bootstrap';
 import { BreadcrumbsItem } from 'react-breadcrumbs-dynamic';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
-import {
-  Bar,
-  CartesianGrid,
-  ComposedChart,
-  Legend,
-  Line,
-  ResponsiveContainer,
-  Tooltip,
-  XAxis,
-  YAxis,
-} from 'recharts';
 
 import { IPagination } from 'model-types';
 import { balanceStatisticPath } from 'routes';
 import { IBalanceStatistic } from 'services/statistic';
-import { formatMoney, ICurrency, number2Money } from 'utils/money';
 
 import { ICurrentOrgIdProps, withCurrentOrgId } from 'components/organizations/current-organization';
 import { SimplePaginator } from 'components/utils/paginator';
-import Tabs from './balance-tabs';
-import Provider from './providers/balance';
+import Chart from './chart';
+import Provider from './provider';
+import Tabs from './tabs';
 
 type IProps = RouteComponentProps<{}> & ICurrentOrgIdProps;
 
-class BalanceChart extends React.PureComponent<IProps> {
-  private valueFormatter = (currency: ICurrency) => {
-    return (value: number) => formatMoney(number2Money(value, currency));
-  }
-
+class BalanceStatisticPage extends React.PureComponent<IProps> {
   private renderPagination = (pagination: IPagination) => {
     return (
       <SimplePaginator
@@ -46,20 +31,7 @@ class BalanceChart extends React.PureComponent<IProps> {
   private renderContent = (balances: IBalanceStatistic, pagination: IPagination | null) => {
     return (
       <>
-        <ResponsiveContainer height={ 300 } width="100%">
-          <ComposedChart
-            data={ balances.data }
-          >
-            <CartesianGrid strokeDasharray="3 3" vertical={ false } />
-            <XAxis dataKey="month"/>
-            <YAxis tickFormatter={ this.valueFormatter(balances.currency) } width={ 100 }/>
-            <Tooltip formatter={ this.valueFormatter(balances.currency) } />
-            <Legend verticalAlign="top" />
-            <Bar dataKey="income" fill="#3366cc" name="Income" />
-            <Bar dataKey="expense" fill="#dc3912" name="Expense" />
-            <Line dataKey="total" stroke="#ff7300" name="Total" />
-          </ComposedChart>
-        </ResponsiveContainer>
+        <Chart balances={ balances } />
         { pagination && this.renderPagination(pagination) }
       </>
     );
@@ -85,4 +57,4 @@ class BalanceChart extends React.PureComponent<IProps> {
   }
 }
 
-export default withRouter(withCurrentOrgId<IProps>(BalanceChart));
+export default withRouter(withCurrentOrgId<IProps>(BalanceStatisticPage));
