--- src/components/statistic/charts.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react';
-
-import Balances from './balance-chart';
-
-const Charts: React.SFC = () => {
-  return (
-    <Balances />
-  );
-};
-
-export default Charts;
